refactor(contracts): split monolithic index schemas into domain modules
